requirements.txt
alembic==1.0.11
Click==7.0
dominate==2.3.5
Flask==1.1.1
Flask-Bootstrap4==4.0.2
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.13
MarkupSafe==1.1.1
python-dateutil==2.8.0
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.5
visitor==0.1.3
Werkzeug==0.15.4
WTForms==2.2.1
